Added queue and tasks

+++ time.c
@@ -1,13 +1,10 @@
 #include "time.h"
 #include <time.h>
 
-uint32_t get_system_time_ms(void) 
-{
-	return (uint32_t)(clock() * 1000 / CLOCKS_PER_SEC);
-}
+uint32_t get_system_time_ms(void) { return (uint32_t)(clock() * 1000 / CLOCKS_PER_SEC); }
 
 void delay_ms(uint32_t ms) 
 {
 	clock_t start = clock();
-	while ((clock() - start) * 1000 / CLOCKS_PER_SEC < ms);
+	while ((uint32_t)((clock() - start) * 1000 / CLOCKS_PER_SEC) < ms);
 }
